requirements.txt
numpy
scipy
scikit-learn
typing
overrides
tqdm
nltk
pytz
tensorboard-pytorch
git+git://github.com/j6mes/drqa@parallel
git+git://github.com/sheffieldnlp/fever-scorer@master
spacy
joblib
pandas
ftfy
allennlp
